+++ backend/canisters/user_index/api/src/updates/update_bot.rs
@@ -1,7 +1,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{OptionUpdate, UserId};
+use types::{BotDefinition, OptionUpdate, UserId};
 
 #[ts_export(user_index, update_bot)]
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -12,6 +12,7 @@ pub struct Args {
     #[ts(as = "types::OptionUpdateString")]
     pub avatar: OptionUpdate<String>, // Image as a data URL
     pub endpoint: Option<String>,
+    pub definition: Option<BotDefinition>,
 }
 
 #[ts_export(user_index, update_bot)]
@@ -29,8 +30,6 @@ pub enum Response {
     OwnerSuspended,
     NewOwnerNotFound,
     NewOwnerSuspended,
-    DefinitionNotFound,
-    DefinitionInvalid,
     DescriptionTooLong,
     TooManyCommands,
 }