+++ frontend/openchat-agent/src/services/icpcoins/icpCoinsClient.ts
@@ -1,12 +1,13 @@
 import type { HttpAgent, Identity } from "@dfinity/agent";
-import type { TokenExchangeRates } from "openchat-shared";
+import type { CryptocurrencyDetails, TokenExchangeRates } from "openchat-shared";
 import { idlFactory, type ICPCoinsService } from "./candid/idl";
 import { CandidService } from "../candidService";
 import { getLatestResponse } from "./mappers";
+import type { ExchangeRateClient } from "../openchatAgent";
 
 const ICPCOINS_CANISTER_ID = "u45jl-liaaa-aaaam-abppa-cai";
 
-export class IcpCoinsClient extends CandidService {
+export class IcpCoinsClient extends CandidService implements ExchangeRateClient {
     private service: ICPCoinsService;
 
     constructor(identity: Identity, agent: HttpAgent) {
@@ -15,7 +16,12 @@ export class IcpCoinsClient extends CandidService {
         this.service = this.createServiceClient<ICPCoinsService>(idlFactory);
     }
 
-    exchangeRates(): Promise<Record<string, TokenExchangeRates>> {
-        return this.handleQueryResponse(() => this.service.get_latest(), getLatestResponse);
+    exchangeRates(
+        supportedTokens: CryptocurrencyDetails[],
+    ): Promise<Record<string, TokenExchangeRates>> {
+        return this.handleQueryResponse(
+            () => this.service.get_latest(),
+            (resp) => getLatestResponse(resp, supportedTokens),
+        );
     }
 }